Move `RedisSentinelCluster` to redis-test.
